OPIK-610 initial structure
