+++ apps/opik-backend/src/main/java/com/comet/opik/domain/TokenUsage.java
@@ -1,0 +1,7 @@
+package com.comet.opik.domain;
+
+import lombok.Builder;
+
+@Builder
+public record TokenUsage(int promptTokens, int completionTokens, int totalTokens) {
+}
