+++ apps/opik-backend/src/main/java/com/comet/opik/domain/llmproviders/LlmProviderStreamHandler.java
@@ -1,0 +1,9 @@
+package com.comet.opik.domain.llmproviders;
+
+import org.glassfish.jersey.server.ChunkedOutput;
+
+public interface LlmProviderStreamHandler {
+    void handleMessage(Object item, ChunkedOutput<String> chunkedOutput);
+    void handleClose(ChunkedOutput<String> chunkedOutput);
+    void handleError(Throwable throwable, ChunkedOutput<String> chunkedOutput);
+}
