+++ apps/opik-backend/src/main/java/com/comet/opik/domain/llmproviders/LlmResponse.java
@@ -1,0 +1,10 @@
+package com.comet.opik.domain.llmproviders;
+
+import com.comet.opik.domain.TokenUsage;
+import lombok.Builder;
+
+import java.util.Map;
+
+@Builder
+public record LlmResponse(String message, TokenUsage tokenUsage, Map<String, Object> metadata) {
+}
